// src/modules/preaching/interfaces/preaching.ts
export interface Preaching {
    id: string;
    userId: string;
    day: string;
    initHour: string;
    finalHour: string;
    createdAt: string;
    updatedAt: string;
}
// ...
export type ReamainingOfHoursRequirement = {
    reamainingOfHoursRequirement: string;
    isNegative: boolean;
}

/**
 * RemainingHoursOfWeeklyRequirement is an object with a remainingHoursOfWeeklyRequirement property that is a string
 * representing the remaining hours requirement.
 *
 * @property {string} remainingHoursOfWeeklyRequirement
 * @property {boolean} isNegative
 */
export type RemainingHoursOfWeeklyRequirement = {
    remainingHoursOfWeeklyRequirement: string;
    isNegative: boolean;
}